fix issue on empty array for clean up inactive users

+++ server/jobs/cleanUpInactiveUsers.js
@@ -10,13 +10,15 @@ exports.cleanUpInactiveUsers = async () => {
         activated: false,
       },
     });
-    inactiveUsers.forEach((user) => inactiveArrayEmails.push(user.email));
-    await User.destroy({ where: { email: inactiveArrayEmails } });
-    logger.log({
-      time: timestamp("YYYY/MM/DD/HH:mm:ss"),
-      level: "info",
-      message: `Users with emails ${inactiveArrayEmails.toString()} was deleted due to inactivated email status`,
-    });
+    if (inactiveUsers.length > 0) {
+      inactiveUsers.forEach((user) => inactiveArrayEmails.push(user.email));
+      await User.destroy({ where: { email: inactiveArrayEmails } });
+      logger.log({
+        time: timestamp("YYYY/MM/DD/HH:mm:ss"),
+        level: "info",
+        message: `Users with emails ${inactiveArrayEmails.toString()} was deleted due to inactivated email status`,
+      });
+    }
   } catch (error) {
     logger.error({
       time: timestamp("YYYY/MM/DD/HH:mm:ss"),
